Update BLS reference tests for new format (#812)
